fix package name error
