+++ soul-sync-data-center/soul-sync-data-websocket/src/main/java/org/dromara/soul/plugin/sync/data/websocket/config/WebsocketConfig.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.dromara.soul.plugin.sync.data.weboscket.config;
+package org.dromara.soul.plugin.sync.data.websocket.config;
 
 import lombok.Data;
